+++ src/main/kotlin/ua/marchenko/artauction/auction/controller/dto/AuctionResponse.kt
@@ -1,17 +1,18 @@
 package ua.marchenko.artauction.auction.controller.dto
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class AuctionResponse(
     val id: String,
     val artworkId: String,
-    val startBid: Double,
+    val startBid: BigDecimal,
     val buyers: List<BidResponse>,
     val startedAt: LocalDateTime,
     val finishedAt: LocalDateTime,
 ) {
     data class BidResponse(
         val buyerId: String,
-        val bid: Double,
+        val bid: BigDecimal,
     )
 }
