+++ src/main/kotlin/ua/marchenko/artauction/auction/controller/dto/CreateAuctionRequest.kt
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.FutureOrPresent
 import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class CreateAuctionRequest(
@@ -12,7 +13,7 @@ data class CreateAuctionRequest(
     val artworkId: String,
 
     @field:Min(value = 1, message = "Auction start bid must be greater than zero")
-    val startBid: Double,
+    val startBid: BigDecimal,
 
     @field:NotNull(message = "Auction start time cannot be null")
     @field:FutureOrPresent(message = "Auction start time must be in the future or present")
